Refactor transforms back into useTemplateStore
