+++ app/(gcforms)/[locale]/(form administration)/form-builder/Start.tsx
@@ -11,8 +11,8 @@ import { safeJSONParse } from "@lib/utils";
 import { FormProperties } from "@lib/types";
 import { validateTemplateSize } from "@lib/utils/validateTemplateSize";
 import { ga } from "@lib/client/clientHelpers";
-import { transformFormProperties } from "@lib/utils/form-builder/transformFormProperties";
 import { validateUniqueQuestionIds } from "@lib/utils/validateUniqueQuestionIds";
+import { transformFormProperties } from "@lib/store/helpers/elements/transformFormProperties";
 
 export const Start = () => {
   const {
@@ -55,10 +55,12 @@ export const Start = () => {
           target.value = "";
           return;
         }
+        // const data = safeJSONParse<FormProperties>(result, cleaner);
+        const data = transformFormProperties(safeJSONParse<FormProperties>(result, cleaner), {
+          cleanRules: false,
+        });
 
-        const data = transformFormProperties(safeJSONParse<FormProperties>(result, cleaner));
-
-        if (!validateUniqueQuestionIds(data.elements)) {
+        if (data && !validateUniqueQuestionIds(data.elements)) {
           setErrors([{ message: t("startErrorDuplicateQuestionId") }]);
           target.value = "";
           return;
